pyproject.toml
[project]
name = "lokikit"
version = "0.1.0"
description = "Minimal CLI to set up and run a local Loki+Promtail+Grafana stack"
authors = [{ name = "Gareth Price" }]
dependencies = [
    "click",
    "pyyaml",
    "loguru",
    "ruff>=0.3.0",
    "basedpyright>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.5.0",
]
requires-python = ">=3.8"

[project.scripts]
lokikit = "lokikit.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[tool.ruff]
target-version = "py311"
line-length = 120

# Allow imports relative to the "lokikit" directory
src = ["lokikit"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "W", "UP", "A", "C4"]
ignore = ["PLR2004"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["lokikit"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
reportImportCycles = "warning"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
reportDuplicateImport = "error"
reportPrivateImportUsage = "warning"
reportOptionalSubscript = "warning"
reportUntypedFunctionDecorator = "warning"
reportConstantRedefinition = "error"
reportUnnecessaryIsInstance = "none"
reportOptionalMemberAccess = "warning"
reportAttributeAccessIssue = "warning"
exclude = ["docs/**", ".venv/**"]
